Move is_valid_head_node() into Dom\Document
