+++ tests/php/test-class-amp-dom-utils.php
@@ -303,112 +303,6 @@ class AMP_DOM_Utils_Test extends WP_UnitTestCase {
 		$this->assertEquals( $body, $output );
 	}
 
-	/**
-	 * Get head node data.
-	 *
-	 * @return array Head node data.
-	 */
-	public function get_head_node_data() {
-		$dom = new Document();
-		return [
-			[
-				AMP_DOM_Utils::create_node( $dom, 'title', [] ),
-				true,
-			],
-			[
-				AMP_DOM_Utils::create_node(
-					$dom,
-					'base',
-					[ 'href' => '/' ]
-				),
-				true,
-			],
-			[
-				AMP_DOM_Utils::create_node(
-					$dom,
-					'script',
-					[ 'src' => 'http://example.com/test.js' ]
-				),
-				true,
-			],
-			[
-				AMP_DOM_Utils::create_node( $dom, 'style', [ 'media' => 'print' ] ),
-				true,
-			],
-			[
-				AMP_DOM_Utils::create_node( $dom, 'noscript', [] ),
-				true,
-			],
-			[
-				AMP_DOM_Utils::create_node(
-					$dom,
-					'link',
-					[
-						'rel'  => 'stylesheet',
-						'href' => 'https://example.com/foo.css',
-					]
-				),
-				true,
-			],
-			[
-				AMP_DOM_Utils::create_node(
-					$dom,
-					'meta',
-					[
-						'name'    => 'foo',
-						'content' => 'https://example.com/foo.css',
-					]
-				),
-				true,
-			],
-			[
-				$dom->createTextNode( " \n\t" ),
-				true,
-			],
-			[
-				$dom->createTextNode( 'no' ),
-				false,
-			],
-			[
-				$dom->createComment( 'hello world' ),
-				true,
-			],
-			[
-				$dom->createProcessingInstruction( 'test' ),
-				false,
-			],
-			[
-				$dom->createCDATASection( 'nope' ),
-				false,
-			],
-			[
-				$dom->createEntityReference( 'bad' ),
-				false,
-			],
-			[
-				$dom->createElementNS( 'http://www.w3.org/2000/svg', 'svg' ),
-				false,
-			],
-			[
-				AMP_DOM_Utils::create_node( $dom, 'span', [] ),
-				false,
-			],
-		];
-	}
-
-	/**
-	 * Test is_valid_head_node().
-	 *
-	 * @dataProvider get_head_node_data
-	 * @covers \AMP_DOM_Utils::is_valid_head_node()
-	 *
-	 * @param DOMNode $node  Node.
-	 * @param bool    $valid Expected valid.
-	 */
-	public function test_is_valid_head_node( $node, $valid ) {
-		$this->assertEquals( $valid, AMP_DOM_Utils::is_valid_head_node( $node ) );
-	}
-
 	public function get_has_class_data() {
 		$dom = new Document();
